requirements.txt
streamlit==1.28.1
pandas==1.5.3
requests==2.31.0
scikit-learn==1.3.0
python-dotenv==1.0.0
pillow==10.0.1